changed int to float

+++ src/generic_utils/script_utils.py
@@ -6,7 +6,7 @@ Date: June 2024
 """
 import numpy as np
 
-def convert_to_array(string, dtype=np.int32):
+def convert_to_array(string, dtype=np.float64):
     """
     Convert a string of comma separated integers to a numpy array.
     """
